requirements.txt
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
numpy==2.2.4
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-telegram-bot==22.0
pytz==2025.1
requests==2.32.3
schedule==1.2.2
six==1.17.0
sniffio==1.3.1
tzdata==2025.1
urllib3==2.3.0
